ISSUE #3652 - refactor revision details file structure. delete redundant components
