+++ frontend/src/v5/ui/components/shared/revisionDetails/components/revisionsListItem/revisionsListItemAuthor/revisionsListItemAuthor.styles.ts
